Change TokenKind::similar_tokens return type
Changes `TokenKind::similar_tokens()` to return a `Vec<TokenKind>`
instead of an `Option<Vec<TokenKind>>`. All callers of the function
were unwrappng this option and calling `contains()` on the `Vec`
inside. The new return type means that `contains()` can be called
directly (with the same results), allowing for cleaner code.